fix chris / daniel concerns
