+++ src/entries/popup/pages/swap/SwapReviewSheet/SwapReviewSheet.tsx
@@ -20,7 +20,7 @@ import { useSwapAssetsToRefreshStore } from '~/core/state/swapAssetsToRefresh';
 import { ParsedSearchAsset } from '~/core/types/assets';
 import { ChainId } from '~/core/types/chains';
 import { truncateAddress } from '~/core/utils/address';
-import { formatCurrency } from '~/core/utils/formatNumber';
+import { processExchangeRateArray } from '~/core/utils/numbers';
 import { isLowerCaseMatch } from '~/core/utils/strings';
 import { isUnwrapEth, isWrapEth } from '~/core/utils/swaps';
 import {
@@ -244,29 +244,8 @@ const SwapReviewSheetWithQuote = ({
   const { minimumReceived, swappingRoute, includedFee, exchangeRate } =
     useSwapReviewDetails({ quote, assetToBuy, assetToSell });
 
-  const cleanNumber = (n: string): number => {
-    return parseFloat(n.replace(/,/g, ''));
-  };
-
-  const formatNumber = (n: string) =>
-    formatCurrency(cleanNumber(n), {
-      notation: 'compact',
-      maximumSignificantDigits: 4,
-    });
-
-  const processArray = (arr: string[]): string[] => {
-    return arr.map((item) => {
-      const parts = item.split(' ');
-      if (parts.length === 5) {
-        const formattedAmount = formatNumber(parts[3]);
-        return `${parts[0]} ${parts[1]} ${parts[2]} ${formattedAmount} ${parts[4]}`;
-      }
-      return item;
-    });
-  };
-
   const formattedExchangeRate = useMemo(
-    () => processArray(exchangeRate),
+    () => processExchangeRateArray(exchangeRate),
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [exchangeRate],
   );
